+++ src/app/shell/layout.tsx
@@ -1,4 +1,4 @@
-import AuthContext from '@/contexts/AuthContext'
+import { AuthContext } from '@/contexts'
 import { PropsWithChildren } from 'react'
 
 export default function ShellLayout({ children }: PropsWithChildren) {
