+++ src/components/Auth/SignupHero.tsx
@@ -1,4 +1,4 @@
-import Review from '@/components/Review'
+import { Review } from '@/components'
 
 const SignupHero = () => {
   return (
