+++ src/components/Forms/SignupForm.tsx
@@ -1,0 +1,5 @@
+const SignupForm = () => {
+  return <h1>Form</h1>
+}
+
+export default SignupForm
