+++ src/components/index.ts
@@ -1,5 +1,14 @@
+// FORMS
+export { default as LoginForm } from './Forms/LoginForm'
+export { default as SignupForm } from './Forms/SignupForm'
+
+// AUTH
+export { default as AuthFormTitle } from './Auth/AuthFormTitle'
+export { default as AuthWrapper } from './Auth/AuthWrapper'
+export { default as SignupHero } from './Auth/SignupHero'
+
+// COMMON
 export { default as Alert } from './Alert'
-export { default as AuthWrapper } from './AuthWrapper'
 export { default as Button } from './Button'
 export { default as CodeBlock } from './CodeBlock'
 export { default as ErrorModal } from './ErrorModal'
@@ -9,6 +18,6 @@ export { default as Input } from './Input'
 export { default as InputError } from './InputError'
 export { default as Label } from './Label'
 export { default as Link } from './Link'
-export { default as LoginForm } from './LoginForm'
 export { default as Logo } from './Logo'
 export { default as Modal } from './Modal'
+export { default as Review } from './Review'
